% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort.R
\name{abort}
\alias{abort}
\title{Abort messenger}
\usage{
abort(message, x, arg = rlang::caller_arg(x), ...)
}
\arguments{
\item{message}{Message to be displayed, formatted via \code{cli::cli_bullets()}. For more information, please read \code{cli::cli_abort}.}

\item{x}{An argument name in the current function.}

\item{arg}{Argument to be traced. When referring to this in argument \code{message}, use \verb{\{.arg \{arg\}\}}.}

\item{...}{Additional items to consider in the message. Any items that uses glue (\code{{}}) must be passed through here.}
}
\value{
An abort message
}
\description{
When an error occurs, it will automatically refer to the function name that the user typed in, instead of the previous function name. Additionally, custom \code{cli} formating can be used to refer different stuff.
This function is based on \code{cli::cli_abort} for formatting. Note that when writing functions, \strong{do not} use pipe (\verb{\%>\%}, \verb{|>}) before this function.

Supports inline-markup themes. For more information on those custom themes, please refer to \code{call::theme} and \code{cli::cli_div}.
}
\examples{
call::sys_abort(message = "test")
}
