% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abort_standard.R
\name{abort_NoArg}
\alias{abort_NoArg}
\alias{abort_WrongClass}
\alias{abort_WrongLength}
\title{Standard abort message}
\usage{
abort_NoArg(x)

abort_WrongClass(x, class)

abort_WrongLength(x, length)
}
\arguments{
\item{x}{The expected argument}

\item{class}{(character) Expected class of the input}

\item{length}{(numeric) Expected length of the input}
}
\value{
A standard abort message
}
\description{
Supported by \code{call::abort}, these provide a standard abort messages to the user and does not require any fiddling.
\emph{Note}: These functions only provide the most basic of input checking. Please double check prior to using it.
}
\examples{
abort_NoArg(x = xlsxFile)
abort_WrongClass(x = xlsxFile, class = "character") #This is for inputing xlsx file to R
abort_WrongLength(x = xlsxFile, length = 1) #Expecting 1 path to R
}
