% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{System Info}
\usage{
info()
}
\value{
A list of system information
}
\description{
The below function tells system on the current system info, including the name of the node, username operating R, and PID.
}
\examples{
call::info()
}
