% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Additional message themeing}
\usage{
theme()
}
\value{
A \code{cli} theme
}
\description{
\code{call::abort} calls this function for additional themes.
}
\examples{
call::theme()
}
